grammar composer: avoid using `->` syntax when function expression has a single primitive argument
